backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
faker>=20.1.0
pillow>=10.1.0
aiofiles>=23.2.1
email-validator>=2.1.0
reportlab>=4.0.7
jinja2>=3.1.2
cryptography>=42.0.0
bcrypt>=4.1.2
fastapi-mail>=1.4.1
PyPDF2>=3.0.1